cleanup -       fixed extra functions on nix-shell -       nix-shell now autocreates virtual env -       do not use caches

+++ default.nix
@@ -1,7 +1,6 @@
 with import <nixpkgs> { };
 pkgs.mkShell rec {
   name = "impurePythonEnv";
-  venvDir = "./.venv";
   buildInputs = [
     python310
   ];
@@ -9,15 +8,20 @@ pkgs.mkShell rec {
   shellHook = ''
     set -h #remove "bash: hash: hashing disabled" warning !
     SOURCE_DATE_EPOCH=$(date +%s)
-
-    if ! [ -d "${venvDir}" ]; then
-      python -m venv "${venvDir}"
-    fi
     export LD_LIBRARY_PATH="${lib.makeLibraryPath [ zlib stdenv.cc.cc ]}":LD_LIBRARY_PATH;
-    source "${venvDir}/bin/activate"
-    python -m pip install --upgrade pip
-    pip install -r requirements.txt
     eval "$extras"
+
+    if ! [ -d .venv ]; then
+      python -m venv .venv
+    fi
+
+    source .venv/bin/activate
+
+    python -m pip install --upgrade pip
+
+    if [ -e requirements.txt ]; then
+      pip install -r requirements.txt
+    fi
   '';
   extras = ''
     pymod() {
